==> 3_prepare_table1.Rmd <==
---
title: "Prepare table 1"
output: 
  html_document:
    toc: true
    keep_md: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r loading packages}
library(tidyverse)
library(survey)
library(plyr)
library(dplyr)
library(factoextra)
library(labelled)
library(tableone)
library(memisc)
library(sjlabelled)
library(mi)
library(boot)
library(glmnet)
library(mice)
library(UpSetR)
library(naniar)
```

# Read in the data

```{r}
folder_path <- "/Users/tamarsofer/Library/CloudStorage/OneDrive-BethIsraelLaheyHealth/Ongoing_papers/2022_gendered_indices"
data_with_sleep <- read_csv(file.path(folder_path, "Data/sol_sofer_sorajja_joon_covariates_20220810.csv"))

# load the prepared dataset
dat <- readRDS(file.path(folder_path, "Data/Data_with_miss.Rds"))

```
# creat table 1

```{r}

vars <-  c("Center", 
           "Age", 
           "Sex",
           "STRAT",
            "Marital_status",
            "Income_level",
            "Employment_status",
            "Occupation",
            "Language_pref",
            "Language_acculturation",
            "Social_acculturation",
            "Ethnic_identity_score", 
            "Current_Health_insurance", 
            "Years_in_US",
            "Background", 
            "Education",
            "STAI10",
            "CESD9",
            "Insomnia",
           "WHIIRS")

survey_obj <- svydesign(id=~PSU_ID, strata=~STRAT, weights=~WEIGHT_FINAL_NORM_OVERALL , data=dat)


tbl1_noweight <- print(CreateTableOne(vars = vars, data =dat, strata = "Sex"), missing=TRUE,varLabels = TRUE,digits =3,pDigits=3,showAllLevels=TRUE)

tbl1_weighted <- print(svyCreateTableOne(vars = vars,  data = survey_obj, strata= "Sex"), missing=TRUE, varLabels = TRUE,digits =3,pDigits=3, showAllLevels=TRUE)


tbl1_w <- tbl1_weighted[,-which(colnames(tbl1_weighted) %in% c("p", "test"))]   
tbl1_unw <- tbl1_noweight[,-which(colnames(tbl1_noweight) %in% c("p", "test"))]

tbl1_comb <- tbl1_unw
col_inds_to_update <- which(colnames(tbl1_unw) %in% c("Female", "Male"))

# update tbl1_comb with the percentages from the weighted table
for (i in col_inds_to_update){
  counts <- sapply(tbl1_unw[,i], function(x){
                  strsplit(x, split = "(", fixed = TRUE)[[1]][1]
                 })
  percent <- sapply(tbl1_w[,i], function(x){
                paste0("(",  strsplit(x, split = "(", fixed = TRUE)[[1]][2])
              })
  tbl1_comb[,i] <- paste0(counts, percent)
}

  
write.csv(tbl1_comb, file.path(folder_path, "Results/Table1_sex_strat.csv"))


```


```{r}
sessionInfo()
```
