==> 6_secondary_train_scores_split.Rmd <==
---
title: "Secondary analysis: independent data split to validate sex associations"
output: 
  html_document:
    toc: true
    keep_md: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r loading packages}
library(tidyverse)
library(survey)
library(plyr)
library(dplyr)
library(factoextra)
library(labelled)
library(memisc)
library(sjlabelled)
library(boot)
library(glmnet)
library(UpSetR)
library(naniar)
library(RColorBrewer)
library(jtools)
library(reshape2)
```
Here we create GISE and GIPSE using ~70% of the data and then evaluate them on the remaining 30% of the data, to check whether the trained indices are still distribute differently between males and females in an independent dataset.

When we split the data, we split by PSU, because individuals within the same PSU may be somewhat correlated to each other. 


# Load the dataset

```{r}
folder_path <- "/Users/tamarsofer/Library/CloudStorage/OneDrive-BethIsraelLaheyHealth/Ongoing_papers/2022_gendered_indices"
data_with_sleep <- read_csv(file.path(folder_path, "Data/sol_sofer_sorajja_joon_covariates_20220810.csv"))

# load the prepared dataset
dat <- readRDS(file.path(folder_path, "Data/Data_with_miss.Rds"))
```

# Train the two indices
Primary genedered index GISE: does not include psychological variables. Secondary gendered index GIPSE: includes also trait anxiety scale and trait depression scale (without the sleep-related item).



```{r}
# Use only complete data in primary analysis
dat_cc <- dat[complete.cases(dat),]

# define train and test indices:
PSU_train <- sample(unique(dat_cc$PSU_ID), 
                    size = ceiling(0.7*length(unique(dat_cc$PSU_ID))))

ID_train <- dat_cc$ID[which(dat_cc$PSU_ID %in% PSU_train)]
ID_test <- setdiff(dat_cc$ID, ID_train)

inds_train <- which(dat_cc$ID %in% ID_train)
inds_test <- which(dat_cc$ID %in% ID_test)

length(ID_train)
length(ID_test)


# define the model matrix for the GISE 
X_gise <- model.matrix(Sex ~
                    Marital_status +
                    Income_level + 
                    Employment_status + 
                    Occupation + 
                    Education + 
                    Current_Health_insurance + 
                    Language_pref + 
                    Language_acculturation + 
                    Social_acculturation + 
                    Ethnic_identity_score + 
                    Years_in_US, dat_cc)[,-1]

X_gipse <- model.matrix(Sex ~
                    Marital_status +
                    Income_level + 
                    Employment_status + 
                    Occupation + 
                    Education + 
                    Current_Health_insurance + 
                    Language_pref + 
                    Language_acculturation + 
                    Social_acculturation + 
                    Ethnic_identity_score + 
                    Years_in_US + 
                    STAI10 + 
                    CESD9, dat_cc)[,-1]

# GISE
set.seed(929)
cv_lasso_gise <- cv.glmnet(X_gise[inds_train,], dat_cc$Sex[inds_train], alpha = 1, family = "binomial", nfolds = 5)

# use the lambda that resulted in minimum CVM to refit the model

set.seed(929)
lasso_model_gise <- glmnet(X_gise[inds_train,], 
                              dat_cc$Sex[inds_train], 
                              family = "binomial", 
                              alpha = 1, 
                              lambda = cv_lasso_gise$lambda.min)
gise_test <- lasso_model_gise %>% predict(newx = X_gise[inds_test,])

# GIPSE
cv_lasso_gipse <- cv.glmnet(X_gipse[inds_train,],
                                dat_cc$Sex[inds_train], 
                                alpha = 1, 
                                family = "binomial", 
                                nfolds = 5)

# use the lambda that resulted in minimum CVM to refit the model

set.seed(929)
lasso_model_gipse <- glmnet(X_gipse[inds_train,], 
                                dat_cc$Sex[inds_train], 
                                family = "binomial", 
                                alpha = 1, 
                                lambda = cv_lasso_gipse$lambda.min)
gipse_test <- lasso_model_gipse %>% predict(newx = X_gipse[inds_test,])

## add to the dataset
dat_test <- dat_cc[inds_test,]
dat_test$GISE  <- gise_test
dat_test$GIPSE  <- gipse_test
```

# Summarize the results


## Creat supplemental figure 1: index distributions by sex
We are not standardizing them at this point, because it does not matter. 

```{r}
dat_test$Sex <- dat_test$Sex
for_plot <- pivot_longer(dat_test, 
                         cols = c("GISE", "GIPSE"), 
                         names_to = "Index_type", 
                         values_to = "Index")

p <- ggplot(for_plot, aes(x = Index)) + 
    geom_histogram(aes(fill = Sex), alpha = 0.6, position = "identity") +
    ggtitle("Histogram of indices stratitifed by sex") + 
    theme_bw() + 
    scale_fill_brewer(palette="Dark2") + 
    xlab("Index value") + ylab("Count") 

p + facet_wrap(~Index_type)

ggsave(file = file.path(folder_path, "Results/Indices_histograms_test_set.pdf"), 
       width = 6, height = 2.7)
```

```{r}
sessionInfo()
```
