==> 7_train_scores_imputed_data.Rmd <==
---
title: "Train gendered indices over the multiply-imputed datasets"
output: 
  html_document:
    toc: true
    keep_md: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r loading packages}
library(tidyverse)
library(survey)
library(plyr)
library(dplyr)
library(factoextra)
library(labelled)
library(memisc)
library(sjlabelled)
library(boot)
library(glmnet)
library(UpSetR)
library(naniar)
library(RColorBrewer)
library(jtools)
library(reshape2)
```

# Load the dataset

```{r}
folder_path <- "/Users/tamarsofer/Library/CloudStorage/OneDrive-BethIsraelLaheyHealth/Ongoing_papers/2022_gendered_indices/"
data_with_sleep <- read_csv(file.path(folder_path, "Data/sol_sofer_sorajja_joon_covariates_20220810.csv"))


# load the prepared dataset
dat <- readRDS(file.path(folder_path, "Data/Data_imputed.Rds"))

# the data with missing values:
# load the prepared dataset
dat_miss <- readRDS(file.path(folder_path, "Data/Data_with_miss.Rds"))

nrow(dat)
```

# Train the two indices
Primary gendered index GISE: does not include psychological variables. Secondary gendered index GIPSE: includes also trait anxiety scale and trait depression scale (without the sleep-related item).



```{r}
vars_gise <- c("Marital_status",
                    "Income_level", 
                    "Employment_status",
                    "Occupation", 
                    "Education", 
                    "Current_Health_insurance", 
                    "Language_pref", 
                    "Language_acculturation", 
                    "Social_acculturation", 
                    "Ethnic_identity_score", 
                    "Years_in_US")
vars_gipse <- c(vars_gise, "STAI10", "CESD9")


for (imp_version in c("", ".1", ".2", ".3", ".4")){
  # define the model matrix for GISE 
  X_gise <- model.matrix(as.formula(paste("Sex ~",
                             paste(vars_gise, collapse = paste0(imp_version,"+")))), dat)
  
  X_gipse <- model.matrix(as.formula(paste("Sex ~",
                             paste(vars_gipse, collapse = paste0(imp_version,"+")))), dat)
                             
        
  # GISE
  set.seed(929)
  cv_lasso_gise <- cv.glmnet(X_gise, dat$Sex, alpha = 1, family = "binomial", nfolds = 5)

  # use the lambda that resulted in minimum CVM to refit the model

  set.seed(929)
  lasso_model_gise <- glmnet(X_gise, dat$Sex, family = "binomial", alpha = 1, lambda = cv_lasso_gise$lambda.min)
  gise <- lasso_model_gise %>% predict(newx = X_gise)

  # GIPSE
  cv_lasso_gipse <- cv.glmnet(X_gipse, dat$Sex, alpha = 1, family = "binomial", nfolds = 5)

# use the lambda that resulted in minimum CVM to refit the model

  set.seed(929)
  lasso_model_gipse <- glmnet(X_gipse, dat$Sex, family = "binomial", alpha = 1, lambda = cv_lasso_gipse$lambda.min)
  gipse <- lasso_model_gipse %>% predict(newx = X_gipse)

## add to the dataset
  dat[[paste0("GISE", imp_version)]]  <- gise
  dat[[paste0("GIPSE", imp_version)]]  <- gipse
             
}



```

# Summarize the results


## Creat supplemental figure 2: average imputed indices, distributions by sex

We are not standardizing them at this point, because it does not matter. 

```{r}
dat$Mean_imputed_gise <- rowMeans(dat[,paste0("GISE", 
                                            c("", ".1", ".2", ".3", ".4"))])

dat$Mean_imputed_gipse <- rowMeans(dat[,paste0("GIPSE", 
                                            c("", ".1", ".2", ".3", ".4"))])
                                            
dat$Sex <- dat$Sex                                            

for_plot <- pivot_longer(dat, 
                         cols = c("Mean_imputed_gise", "Mean_imputed_gipse"), 
                         names_to = "Index_type", 
                         values_to = "Index")

p <- ggplot(for_plot, aes(x = Index)) + 
    geom_histogram(aes(fill = Sex), alpha = 0.6, position = "identity") +
    ggtitle("Histogram of indices stratified by sex") + 
    theme_bw() + 
    scale_fill_brewer(palette="Dark2") + 
    xlab("Index value") + ylab("Count") 

p + facet_wrap(~Index_type)

ggsave(file = file.path(folder_path, "Results/Indices_imputed_histograms.pdf"), 
       width = 6, height = 2.7)
```

# Make a figure based only on imputed observations:

```{r}
ID_imp <- dat_miss$ID[!complete.cases(dat_miss)]
length(ID_imp)

p <- ggplot(for_plot[for_plot$ID %in% ID_imp,], aes(x = Index)) + 
    geom_histogram(aes(fill = Sex), alpha = 0.6, position = "identity") +
    ggtitle("Histogram of indices stratified by sex") + 
    theme_bw() + 
    scale_fill_brewer(palette="Dark2") + 
    xlab("Index value") + ylab("Count") 

p + facet_wrap(~Index_type)

ggsave(file = file.path(folder_path, "Results/Indices_only_imputed_histograms.pdf"), 
       width = 6, height = 2.7)
```


# final figure: stratified by age. 

```{r}

dat$Age_group <- as.factor(cut(dat$Age, 
                            breaks =  c(min(dat$Age), 30,40, 50, 60, max(dat$Age)), 
                            include.lowest = TRUE))
                            
                            
for_plot <- pivot_longer(dat, 
                         cols = c("Mean_imputed_gise", "Mean_imputed_gipse"), 
                         names_to = "Index_type", 
                         values_to = "Index")

index_medians <- for_plot %>% group_by(Index_type, Age_group, Sex) %>% 
  summarize_at(vars(Index), list(index_median = median)) 


p <- ggplot(for_plot, aes(x = Index)) + 
  geom_histogram(aes(fill = Sex), alpha = 0.6, position = "identity") +
  ggtitle("Histogram of indices stratitifed by sex") + 
  theme_bw() + 
  scale_fill_brewer(palette="Dark2") + 
  xlab("Index value") + ylab("Count") 

p + facet_grid(Age_group~Index_type) + 
  geom_vline(data = index_medians, mapping = aes(xintercept = index_median, color = Sex)) + 
  scale_color_brewer(palette="Dark2")

ggsave(file = file.path(folder_path, "Results/Indices_imputed_histograms_by_age.pdf"), 
       width = 6, height = 6)
```


# Save the dataset for future use

```{r}
saveRDS(dat, file.path(folder_path, "Data/Data_imputed_with_indices.Rds"))

```

```{r}
sessionInfo()
```
