==> 11_sex_indices_interaction_models.Rmd <==
---
title: "Regression of WHIIRS and insomnia with sex, GISE/GIPSE, and their interaction"
output: 
  html_document:
    toc: true
    keep_md: true
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r loading packages}
library(tidyverse)
library(survey)
library(plyr)
library(dplyr)
library(factoextra)
library(labelled)
library(memisc)
library(sjlabelled)
library(boot)
library(glmnet)
library(UpSetR)
library(naniar)
library(RColorBrewer)
library(jtools)
library(reshape2)
# source p-value formatting function
source("format_pvalue.R")
```

# Load and prepare the dataset

```{r}
folder_path <- "/Users/tamarsofer/Library/CloudStorage/OneDrive-BethIsraelLaheyHealth/Ongoing_papers/2022_gendered_indices"
dat_no_indices <- readRDS(file.path(folder_path, "Data/Data_with_miss.Rds"))

# the data with missing values:
# load the prepared dataset with indices
dat_miss <- readRDS(file.path(folder_path, "Data/Data_complete_case_with_indices.Rds"))


# load the imputed dataset with indices
dat_imp <- readRDS(file.path(folder_path, "Data/Data_imputed_with_indices.Rds"))

print(paste("The complete case dataset has", nrow(dat_miss), "individuals"))
print(paste("The imputed dataset has", nrow(dat_imp), "individuals"))

# because of the survey sampling design, we need to "merge" 
# the complete case dataset with design variables of the complete dataset
dat <- merge(dat_miss, 
             dat_no_indices[,c("ID", "STRAT" ,"PSU_ID", "WEIGHT_FINAL_NORM_OVERALL")], 
             by = c("ID", "STRAT" ,"PSU_ID", "WEIGHT_FINAL_NORM_OVERALL"), all = TRUE)

dat_imp <- merge(dat_imp, 
                 dat_no_indices[,c("ID", "WEIGHT_FINAL_NORM_OVERALL", "PSU_ID")], 
                 by = c("ID"))

nrow(dat) == nrow(dat_imp)
```
# add scaled indices to describe effect per 1 SD increase

```{r}
survey_cc <- svydesign(id = ~PSU_ID,
                          strata = ~STRAT,
                          weights = ~WEIGHT_FINAL_NORM_OVERALL,
                          nest = TRUE,
                          data = dat)
survey_cc <- subset(survey_cc, complete.cases(dat))

survey_imp <- svydesign(id = ~PSU_ID,
                          strata = ~STRAT,
                          weights = ~WEIGHT_FINAL_NORM_OVERALL,
                          nest = TRUE,
                          data = dat_imp)

cur_mean <- svymean(~GISE, survey_cc)[1]
cur_sd <- sqrt(svyvar(~GISE, survey_cc)[1])
dat_miss$GISE_scaled <- (dat_miss$GISE - cur_mean)/cur_sd

cur_mean <- svymean(~GIPSE, survey_cc)[1]
cur_sd <- sqrt(svyvar(~GIPSE, survey_cc)[1])
dat_miss$GIPSE_scaled <- (dat_miss$GIPSE - cur_mean)/cur_sd

for (suffix in c("", ".1", ".2", ".3", ".4")){
  cur_mean <- svymean(as.formula(paste0("~GISE", suffix)), survey_imp)[1]
  cur_sd <- sqrt(svyvar(as.formula(paste0("~GISE", suffix)), survey_imp)[1])
  dat_imp[[paste0("GISE_scaled", suffix)]] <- (dat_imp[[paste0("GISE", suffix)]] - cur_mean)/cur_sd

  cur_mean <- svymean(as.formula(paste0("~GIPSE", suffix)), survey_imp)[1]
  cur_sd <- sqrt(svyvar(as.formula(paste0("~GIPSE", suffix)), survey_imp)[1])
  dat_imp[[paste0("GIPSE_scaled", suffix)]] <- (dat_imp[[paste0("GIPSE", suffix)]] - cur_mean)/cur_sd
}


# recompute the survey objects:
dat <- merge(dat_miss, 
             dat_no_indices[,c("ID", "STRAT" ,"PSU_ID", "WEIGHT_FINAL_NORM_OVERALL")], 
             by = c("ID", "STRAT" ,"PSU_ID", "WEIGHT_FINAL_NORM_OVERALL"), all = TRUE)

survey_cc <- svydesign(id = ~PSU_ID,
                          strata = ~STRAT,
                          weights = ~WEIGHT_FINAL_NORM_OVERALL,
                          nest = TRUE,
                          data = dat)
survey_cc <- subset(survey_cc, complete.cases(dat))

survey_imp <- svydesign(id = ~PSU_ID,
                          strata = ~STRAT,
                          weights = ~WEIGHT_FINAL_NORM_OVERALL,
                          nest = TRUE,
                          data = dat_imp)


```
# Prepare model variables

```{r}
model1 <- c("Age", "Center", "Background", "Sex*GISE_scaled")
model2 <- c("Age", "Center", "Background", "Sex*GIPSE_scaled")


```

#  Evaluate models and extract results
```{r}

# function to extract the estimated effect of an exposure on a binary outcome
extract_one_exp_logistic <- function(mod, exposure, round_digit = 2){
   est <- summary(mod)$coef
   confint <- confint(mod)
  
  ind <- which(exposure ==  rownames(est))
  
  df_out <- data.frame(exposure = exposure, 
                       est = round(exp(est[ind, "Estimate"]),round_digit), 
                       CI = paste0("(", round(exp(confint[ind, 1]), round_digit),
                                   ",",round(exp(confint[ind, 2]), round_digit) , ")"),
                       pval = formatC(est[ind, "Pr(>|t|)"], digits = round_digit, format = "E"), 
                       paper_pval = format_pvalue(est[ind, "Pr(>|t|)"]))
 

  return(df_out)
}



extract_one_exp_linear <- function(mod, exposure, round_digit = 2){
   est <- summary(mod)$coef
   confint <- confint(mod)
  
  ind <- which(exposure ==  rownames(est))
  
  df_out <- data.frame(exposure = exposure, 
                       est = round(est[ind, "Estimate"],round_digit), 
                       CI = paste0("(", round(confint[ind, 1], round_digit),
                                   ",",round(confint[ind, 2], round_digit) , ")"),
                       pval = formatC(est[ind, "Pr(>|t|)"], digits = round_digit, format = "E"), 
                       paper_pval = format_pvalue(est[ind, "Pr(>|t|)"]))
 

  return(df_out)
}

extract_one_exp <- function(mod, exposure, round_digit = 2, model_name){
  if (length(grep("insomnia", model_name)) == 1) extract_one_exp_logistic(mod, exposure, round_digit = 2) else
  if (length(grep("whiirs", model_name)) == 1) extract_one_exp_linear(mod, exposure, round_digit = 2)
}


table6 <- data.frame(row.names = c("Insomnia_with_GISE",
                                   "Insomnia_with_GIPSE",
                                   "WHIIRS_with_GISE",
                                   "WHIIRS_with_GIPSE"),
                     male_eff = rep(NA, 4),
                     male_CI = NA,
                     male_pval = NA,
                     male_paper_pval = NA,
                     index_eff = NA,
                     index_CI = NA,
                     index_pval = NA,
                     index_paper_pval = NA,
                     interaction_eff = NA,
                     interaction_CI = NA,
                     interaction_pval = NA,
                     interaction_paper_pval = NA)

# fit models (sex combined)
model1_fit_insomnia <- svyglm(formula = as.formula(paste0("Insomnia~", paste(model1, collapse = "+"))), design = survey_cc, family = quasibinomial())

model2_fit_insomnia <- svyglm(formula = as.formula(paste0("Insomnia~", paste(model2, collapse = "+"))), design = survey_cc, family = quasibinomial())

model1_fit_whiirs <- svyglm(formula = as.formula(paste0("WHIIRS~", paste(model1, collapse = "+"))), design = survey_cc, family = gaussian())

model2_fit_whiirs <- svyglm(formula = as.formula(paste0("WHIIRS~", paste(model2, collapse = "+"))), design = survey_cc, family = gaussian())


row_model <- data.frame(row_ind = 1:4, row_name = row.names(table6), 
                        model_name = c("model1_fit_insomnia", 
                                       "model2_fit_insomnia",
                                       "model1_fit_whiirs", 
                                       "model2_fit_whiirs"), 
                        index_name = rep(c("GISE_scaled", "GIPSE_scaled"),2))

for (i in 1:4){
  table6[i,
       c("index_eff", "index_CI",  "index_pval", "index_paper_pval")]  <- 
          extract_one_exp(get(paste0(row_model$model_name[i])) , 
                                 exposure = row_model$index_name[i],
                            model_name = row_model$model_name[i])[c("est", "CI", "pval", "paper_pval")]
  
    table6[i,
       c("male_eff", "male_CI",  "male_pval", "male_paper_pval")]  <- 
          extract_one_exp(get(paste0(row_model$model_name[i])),
                            model_name = row_model$model_name[i] , 
                                 exposure = "SexMale")[c("est", "CI", "pval", "paper_pval")]
    
      table6[i,
       c("interaction_eff", "interaction_CI",  "interaction_pval", "interaction_paper_pval")]  <- 
          extract_one_exp(get(paste0(row_model$model_name[i])),
                            model_name = row_model$model_name[i], 
                                 exposure = paste0("SexMale:", row_model$index_name[i]))[c("est", "CI", "pval","paper_pval")]
  
}



write.csv(table6, file = file.path(folder_path, "Results/Table6_interactions.csv"))

```


# Table 6 using imputed dataset (Supplementary Table ??)

Functions that will help us
```{r}
# can't use the pool function from the MI package
# instead write a function that fit a model on the 5 datasets, and applies Rubin's rule.
# for simplicity, assume WHIIRS is outcome, assume 5 imputations

rubins_rule <- function(ests, ses, Insomnia = TRUE, round_digit = 2){
  vars <- ses^2
  m <- length(ses)
  
  pooled_est <- mean(ests)
  within_imp_var <- mean(vars)
  between_imp_var <- sum((vars - within_imp_var)^2)/(m -1)
  
  pooled_var <- within_imp_var + (1+1/m)*between_imp_var
  pooled_se <- sqrt(pooled_var)
  
  F_test_stat <- pooled_est^2/pooled_var
  F_df1 <- 1
  r <- (1 + 1/m)*between_imp_var/pooled_est
  F_df2 <- (m - 1)*(1 + 1/r)^2
  
  F_pval <- pf(F_test_stat, F_df1, F_df2, lower.tail = FALSE)
  
  if (Insomnia){
      CI <- paste0("(", round(exp(pooled_est - 1.96*pooled_se), round_digit), ",", round(exp(pooled_est + 1.96*pooled_se), round_digit), ")")
  
  return(data.frame(est = round(exp(pooled_est), round_digit), CI = CI, pval = F_pval))
  } else{
      CI <- paste0("(", round(pooled_est - 1.96*pooled_se, round_digit), ",", round(pooled_est + 1.96*pooled_se, round_digit), ")")
  
  return(data.frame(est = round(pooled_est, round_digit), CI = CI, pval = F_pval))
  }

  
}


extract_one_exp_mi <- function(survey_imp,  model_vars, exposure, 
                               round_digit = 2, Insomnia = TRUE){
  
  imp_suffix <-  c("", ".1", ".2", ".3", ".4")
  exposure_no_suffix <- exposure
  
  est <- se <- rep(NA, 5)
  names(est) <- names(se) <-  imp_suffix
  for (i in 1:length(imp_suffix)){
    suffix <- imp_suffix[i]
    
    if (exposure == "SexMale"){ 
      exposure <- exposure_no_suffix 
    } else{
      exposure <- paste0(exposure_no_suffix, suffix)
    } 
    
    if (Insomnia){
          cur_fit <- svyglm(formula = as.formula(paste0("Insomnia", suffix, "~", paste(paste0(model_vars, suffix), collapse = "+"))), design = survey_imp, family = quasibinomial())
    } else{
          cur_fit <- svyglm(formula = as.formula(paste0("WHIIRS", suffix, "~", paste(paste0(model_vars, suffix), collapse = "+"))), design = survey_imp, family = gaussian())
    }
    

    
    fit_est <- summary(cur_fit)$coef
    ind_exp <- which(exposure ==  rownames(fit_est))
    est[i] <- fit_est[ind_exp, "Estimate"]
    se[i] <- fit_est[ind_exp, "Std. Error"]
    
  }
  
  # now apply Rubin's rule
  
  res <- rubins_rule(est, se, round_digit = round_digit, Insomnia)

  
  df_out <- data.frame(exposure = exposure, 
                       est = res[["est"]], 
                       CI = res[["CI"]], 
                       pval = formatC(res[["pval"]], digits = round_digit, format = "E"), 
                       paper_pval = format_pvalue(res[["pval"]]))
 

  return(df_out)
}



```




And prepare the table using imputed data:
```{r}


table6_imp <- data.frame(row.names = c("Insomnia_with_GISE",
                                   "Insomnia_with_GIPSE",
                                   "WHIIRS_with_GISE",
                                   "WHIIRS_with_GIPSE"),
                     male_eff = rep(NA, 4),
                     male_CI = NA,
                     male_pval = NA,
                     male_paper_pval = NA,
                     index_eff = NA,
                     index_CI = NA,
                     index_pval = NA,
                     index_paper_pval = NA,
                     interaction_eff = NA,
                     interaction_CI = NA,
                      interaction_pval = NA, 
                     interaction_paper_pval = NA)



row_model <- data.frame(row_ind = 1:4, row_name = row.names(table6), 
                        model_name = c("model1_fit_insomnia", 
                                       "model2_fit_insomnia",
                                       "model1_fit_whiirs", 
                                       "model2_fit_whiirs"), 
                        index_name = rep(c("GISE_scaled", "GIPSE_scaled"),2))

for (i in 1:4){
  
  if (i %in% c(1,3)){
    model_vars <- model1
  } else {
    model_vars <- model2
  }
  
  table6_imp[i,
       c("index_eff", "index_CI",  "index_pval", "index_paper_pval")]  <- 
          extract_one_exp_mi(survey_imp, 
              model_vars = model_vars, 
              exposure = row_model$index_name[i], 
               Insomnia = ifelse(i %in% c(1,2), TRUE, FALSE))[, c("est", "CI", "pval", "paper_pval")]
  
  
        
    table6_imp[i,
       c("male_eff", "male_CI",  "male_pval", "male_paper_pval")]  <- 
              extract_one_exp_mi(survey_imp, 
                    model_vars = model_vars, 
                     exposure =  "SexMale", 
                     Insomnia = ifelse(i %in% c(1,2), TRUE, FALSE))[, c("est", "CI", "pval", "paper_pval")]
    
    
     table6_imp[i, c("interaction_eff", "interaction_CI",  "interaction_pval", "interaction_paper_pval")]  <- 
                extract_one_exp_mi(survey_imp, 
                model_vars = model_vars, 
                exposure = paste0("SexMale:", row_model$index_name[i]), 
                Insomnia = ifelse(i %in% c(1,2), TRUE, FALSE))[, c("est", "CI", "pval", "paper_pval")]
        
   
}





write.csv(table6_imp, file = file.path(folder_path, "Results/Supp_table_6_with_mi.csv"))

```

```{r}
sessionInfo()
```
